MnemoCore Chatbot/requirements.txt

streamlit==1.35.0
openvino==2025.2.0
transformers==4.40.2
optimum==1.17.1
torch==2.1.2
sentence-transformers==2.2.2

langchain==0.1.16
langchain-community==0.0.36
langchain-core==0.1.50

chromadb==0.4.24
faiss-cpu==1.7.4
tqdm==4.66.2

pypdf==3.17.4
unstructured==0.13.5

faster-whisper==0.10.0
ffmpeg-python==0.2.0
soundfile==0.12.1

python-dotenv==1.0.1
Pillow==10.3.0
pydantic==2.6.4

streamlit-extras==0.3.6
